feat: eslint & husky update files npm i ok storybook without stories ok rollup ok

+++ scripts/remove-published-version.js
@@ -1,33 +1,33 @@
-const inquirer = require("inquirer");
-const { execSync } = require("child_process");
+const inquirer = require('inquirer');
+const { execSync } = require('child_process');
 
 const question = [
   {
-    name: "version",
+    name: 'version',
     message:
       "Which version (format X.X.X) of 'roket-ui' package would you like to unpublish ?",
   },
 ];
 
-inquirer.prompt(question).then((answers) => {
+inquirer.prompt(question).then(answers => {
   const isValid = RegExp(/\d*\.\d*\.\d*/gm).exec(answers.version);
   if (!isValid) {
     return console.log(
-      "Wrong format, version must be x.x.x for example 1.20.3"
+      'Wrong format, version must be x.x.x for example 1.20.3'
     );
   }
   inquirer
     .prompt([
       {
-        name: "confirm",
+        name: 'confirm',
         message: `Do you really want to unpublish rocket-ui-dev-version@${answers.version} ? `,
-        type: "confirm",
+        type: 'confirm',
       },
     ])
-    .then((confirmAnswers) => {
+    .then(confirmAnswers => {
       if (confirmAnswers.confirm) {
         execSync(`npm unpublish rocket-ui-dev-version@${answers.version}`, {
-          stdio: "inherit",
+          stdio: 'inherit',
         });
       }
     });
